+++ packages/sdocs/src/lib/ui/Icon/Icon.svelte
@@ -1,8 +1,12 @@
 <script lang="ts">
+	import accessibilitySvg from './icons/accessibility.svg?raw';
 	import bookmarkSvg from './icons/bookmark.svg?raw';
+	import bugSvg from './icons/bug.svg?raw';
 	import chevronDownSvg from './icons/chevron-down.svg?raw';
 	import circleAlertSvg from './icons/circle-alert.svg?raw';
 	import circleCheckSvg from './icons/circle-check.svg?raw';
+	import circleDashedSvg from './icons/circle-dashed.svg?raw';
+	import circleDotSvg from './icons/circle-dot.svg?raw';
 	import chevronLeftSvg from './icons/chevron-left.svg?raw';
 	import chevronRightSvg from './icons/chevron-right.svg?raw';
 	import chevronUpSvg from './icons/chevron-up.svg?raw';
@@ -12,8 +16,11 @@
 	import databaseSvg from './icons/database.svg?raw';
 	import diamondSvg from './icons/diamond.svg?raw';
 	import externalLinkSvg from './icons/external-link.svg?raw';
+	import eyeSvg from './icons/eye.svg?raw';
+	import flaskSvg from './icons/flask.svg?raw';
 	import fileCodeSvg from './icons/file-code.svg?raw';
 	import infoSvg from './icons/info.svg?raw';
+	import lightbulbSvg from './icons/lightbulb.svg?raw';
 	import fileTextSvg from './icons/file-text.svg?raw';
 	import folderOpenSvg from './icons/folder-open.svg?raw';
 	import folderSvg from './icons/folder.svg?raw';
@@ -29,10 +36,14 @@
 	import zapSvg from './icons/zap.svg?raw';
 
 	const icons: Record<string, string> = {
+		'accessibility': accessibilitySvg,
 		'bookmark': bookmarkSvg,
+		'bug': bugSvg,
 		'chevron-down': chevronDownSvg,
 		'circle-alert': circleAlertSvg,
 		'circle-check': circleCheckSvg,
+		'circle-dashed': circleDashedSvg,
+		'circle-dot': circleDotSvg,
 		'chevron-left': chevronLeftSvg,
 		'chevron-right': chevronRightSvg,
 		'chevron-up': chevronUpSvg,
@@ -42,8 +53,11 @@
 		'database': databaseSvg,
 		'diamond': diamondSvg,
 		'external-link': externalLinkSvg,
+		'eye': eyeSvg,
+		'flask': flaskSvg,
 		'file-code': fileCodeSvg,
 		'info': infoSvg,
+		'lightbulb': lightbulbSvg,
 		'file-text': fileTextSvg,
 		'folder-open': folderOpenSvg,
 		'folder': folderSvg,